refactor: update sc2utils package (#3)
* refactor: Recode Arg Parser

* refactor: Remove Property Reader

Property reader has been replaced with ArgParser as it can read basic ini files which covers the ExecuteInfo file format

* refactor: remove dirent.h

replaced dirent.h and its usages with the more cross-platform filesystem header provided in C++17

* refactor: delete unused header

simple_serialization was not used anywhere in the project. This removes that header. Will readd or replace if found to need it later but proper serialization implementation by calling program would be better